requirements.txt
discord.py
pandas
openpyxl
schedule
google-auth
google-auth-oauthlib
google-api-python-client
aiohttp